Dutch localization for natural time
This commit provides a localized version of the natural time base class for
both the nl-NL and nl-BE locale. In general, I think there is some room for
grouping region-specific variants. For instance, fr-BE can probably be identical
to fr-FR and fr-CA as far as time is concerned, and nl-BE and fr-BE should
share public holidays.