Ascension day is also a holiday in NL, bugfix for same-day holidays
After adding ascension day as a national public holiday for the Netherlands,
it turned out that in 2005, ascension day was on the same day as liberation day.
This commit introduces a general fix for the exception that occurred in that
situation by picking one of the holidays, if both are observed on the same day.

Please also see commit fc6e331 in the multipleObservances branch which is kind of
related, but different (in NL we don't get to reuse a double holiday on some
other day).

Perhaps the BuildObservancesMap should be overridden in the concrete class should be
overrided instead of the change to the base class, but I felt this was an issue
that is more general than just the dutch case.