Update EN_ZAHolidayStrategy.cs
Error occurs where the key already exists, this happens with the 26th December, as the 25th in on a Sunday, the holiday if forfeited if the next Day is also a public holiday as in this case. SO we need to check if this occurs and if it does we just update the value of the key to the correct name. This was you will not end up with two Christmas's etc...